+++ CHANGELOG.md
@@ -19,7 +19,7 @@ We use the following categories for changes:
 - Added `prom_api.get_default_chunk_interval()` and `prom_api.get_metric_chunk_interval(TEXT)` [#464].
 
 ### Fixed
-- Incorrect type coercion when using `tag_map` with `=` operator [#345]
+- Incorrect type coercion when using `tag_map` with `=` operator [#462]
 
 ### Changed